Create Token Extraction Helper Function
